src/main/java/com/udemy/gestioninfraestructuraapi/model/Servidor.java
package com.udemy.gestioninfraestructuraapi.model;

public class Servidor {

    private Long codigo;
    private String nombre;
    private String ip;
    private String os;

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getIp() {
        return ip;
    }

    public void setIp(String ip) {
        this.ip = ip;
    }

    public String getOs() {
        return os;
    }

    public void setOs(String os) {
        this.os = os;
    }

    public Long getCodigo() {
        return codigo;
    }

    public void setCodigo(Long codigo) {
        this.codigo = codigo;
    }

}
